[FIX] 코드 분리
